requirements.txt
pandas==0.23.4
requests==2.21.0
matplotlib==3.0.2
numpy==1.15.4
Pillow==5.4.1
